pyproject.toml
[project]
name = "browser-conf"
version = "0.1.0"
description = "Declarative configuration management for Zen Browser"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.scripts]
zen-apply = "browser_conf.apply:main"
json-to-yaml = "browser_conf.json_to_yaml:json_to_yaml"

[tool.uv]
package = true

[tool.setuptools]
packages = ["browser_conf"]
